Replace svg surface id's with 0000
SVG's produced by R's svg device seem to include a random surface id number for some group of elements. Not sure what the point of this is. But annoyingly, Rstudio's knit button and the rmarkdown::render command cannot seem to agree on a cache, so this random ID is always changing, so the figures always show up as modified in git, despite no substantive changes. I put in a git clean filter to find and replace these surface ids with 0000 in every case. Hope that's fine.