Set a fixed width for R output in several labs and slide decks.
Rmarkdown can't seem to make up it's mind about when to insert a line break in it's output. Sometimes it wraps the same output after 80, or 85, or 90, without anything changing about the code itself, just re-rendering the output. Annoyingly, this blows up git diff with meaningless changes. I think it's another case of the knit button and render failing to agree on global chunk options or a cache, and the output then fluctuates ever so slightly depending on who is doing the rendering. So, I just set a fixed console width for a few chunks where this seemed to be a consistent issue.